Remove omitted filenames from chunk.files
When a file is omitted from `compilation.assets`, its filename still
referenced by the chunk it belongs to. This can confuse other webpack
plugins that reference `chunk.files`, such as html-webpack-plugin,
which uses `chunk.files` to determine which <link> and <script> tags
to include in an html file.

This commit removes the omitted filename from the chunk it belongs to,
in addition to removing it from the compilation's assets.